chore: expand to monorepo with workspace packages
- Rename shared/ → packages/shared/ as @showtracker/shared
- Add packages/api-client/ as @showtracker/api-client (framework-agnostic
  API client with injected token getter and configurable base URL)
- Add npm workspaces ["packages/*", "apps/*"] to root package.json
- Update tsconfig.json and vite.config.ts path aliases for new locations
- Update vercel.json includeFiles to reference packages/shared/**
- Add turbo.json for build pipeline
