+++ client/src/lib/queryClient.ts
@@ -1,70 +1,14 @@
-import { QueryClient, QueryFunction } from "@tanstack/react-query"
+import {
+  makeQueryClient,
+  setApiBaseUrl,
+  setApiTokenGetter,
+  apiRequest,
+  getQueryFn,
+} from "@showtracker/api-client"
 
-let getApiToken: () => Promise<string | null> = async () => null
+export { setApiTokenGetter, apiRequest, getQueryFn, setApiBaseUrl }
 
-export function setApiTokenGetter(fn: () => Promise<string | null>) {
-  getApiToken = fn
-}
+// Web app uses relative URLs
+setApiBaseUrl("")
 
-async function throwIfResNotOk(res: Response) {
-  if (!res.ok) {
-    const text = (await res.text()) || res.statusText
-    throw new Error(`${res.status}: ${text}`)
-  }
-}
-
-export async function apiRequest(
-  method: string,
-  url: string,
-  data?: unknown | undefined
-): Promise<Response> {
-  const token = await getApiToken()
-  const headers: Record<string, string> = {
-    ...(data ? { "Content-Type": "application/json" } : {}),
-    ...(token ? { Authorization: `Bearer ${token}` } : {}),
-  }
-  const res = await fetch(url, {
-    method,
-    headers,
-    body: data ? JSON.stringify(data) : undefined,
-    credentials: "include",
-  })
-
-  await throwIfResNotOk(res)
-  return res
-}
-
-type UnauthorizedBehavior = "returnNull" | "throw"
-export const getQueryFn: <T>(options: {
-  on401: UnauthorizedBehavior
-}) => QueryFunction<T> =
-  ({ on401: unauthorizedBehavior }) =>
-  async ({ queryKey }) => {
-    const token = await getApiToken()
-    const res = await fetch(queryKey.join("/") as string, {
-      credentials: "include",
-      headers: token ? { Authorization: `Bearer ${token}` } : {},
-    })
-
-    if (unauthorizedBehavior === "returnNull" && res.status === 401) {
-      return null
-    }
-
-    await throwIfResNotOk(res)
-    return await res.json()
-  }
-
-export const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      queryFn: getQueryFn({ on401: "throw" }),
-      refetchInterval: false,
-      refetchOnWindowFocus: false,
-      staleTime: Infinity,
-      retry: false,
-    },
-    mutations: {
-      retry: false,
-    },
-  },
-})
+export const queryClient = makeQueryClient()
