+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "type": "module",
   "license": "MIT",
+  "workspaces": ["packages/*", "apps/*"],
   "scripts": {
     "dev": "NODE_ENV=development tsx api/index.ts",
     "build": "vite build",
