+++ packages/api-client/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@showtracker/api-client",
+  "version": "1.0.0",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "peerDependencies": {
+    "@tanstack/react-query": "^5"
+  }
+}
