--- shared/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "commonjs"
-}
