+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["client/src/**/*", "shared/**/*", "server/**/*", "api/**/*", "dev/**/*"],
+  "include": ["client/src/**/*", "packages/shared/**/*", "server/**/*", "api/**/*", "dev/**/*"],
   "exclude": ["node_modules", "build", "dist", "**/*.test.ts"],
   "compilerOptions": {
     "incremental": true,
@@ -17,7 +17,9 @@
     "types": ["node", "vite/client"],
     "paths": {
       "@/*": ["./client/src/*"],
-      "@shared/*": ["./shared/*"]
+      "@shared/*": ["./packages/shared/*"],
+      "@showtracker/shared": ["./packages/shared/schema.ts"],
+      "@showtracker/api-client": ["./packages/api-client/src/index.ts"]
     }
   }
 }
