+++ vercel.json
@@ -6,7 +6,7 @@
     "api/index.ts": {
       "memory": 1024,
       "maxDuration": 30,
-      "includeFiles": "server/**,shared/**"
+      "includeFiles": "server/**,packages/shared/**"
     }
   },
   "rewrites": [
